CartoonGAN/requirements.txt
gradio==3.0.5
huggingface-hub==0.6.0
pillow==9.1.1
torch==1.11.0
torchvision==0.12.0